Added #defines for XP3i project
The changes made here will work for the XP3I project. They should not effect any other project made  with the libexpat library.

To add the library to a new project add the project name as a macro to the project then guard all the needed flags, such as XML_CONTEXT_BYTES, with #ifdef [project name].